requirements.txt
boto3
mysql-connector-python==8.0.11

